fix: clear 31 of 32 compiler warnings, and two real defects behind them
The host test build carried 110 warnings across 32 unique sites. Most were
noise, but reading each one rather than blanket-suppressing found two things
worth having:

- TextBlock.cpp: file.read() returns a signed count and -1 on error, compared
  directly against a size_t. The -1 promoted to a huge unsigned value, so the
  error branch was taken for the right reason by accident. Now explicit.
- ClaudeChatActivity/NoteEditorActivity carried a stale 'Reserve the right
  gutter' comment directly contradicting the one below it -- left behind by
  1c795eeca this morning when the side-button hints and their gutter were
  removed. Both removed, plus a dead contentBottom local.

Format specifiers: FontDecompressor printed uint32_t with %lu (wrong on a
64-bit host, right on device by coincidence) and BaseTheme printed size_t with
%u. Now %u and %zu respectively.

Sign compares: ParsedText follows the static_cast<size_t>(i) idiom the same
loop already used three lines up.

The rest are unused parameters on virtual overrides and template
specialisations, unnamed rather than deleted so the signatures still match.
Two dead file-local constants and one dead local removed.

Note bufferRestored: a bool& out-parameter neither override writes and the
caller never reads back. Unnamed here; removing it from the virtual signature
is a refactor, not a warning fix.

One warning left, deliberately: KeyboardIconTarget.h:85 is a private field
only live under CROSSPOINT_RENDER_SCALE > 1.

290/290 tests pass. -e default and -e simulator build from a private cache.

+++ lib/Epub/Epub/ParsedText.cpp
@@ -760,10 +760,10 @@ std::vector<size_t> ParsedText::computeLineBreaks(const GfxRenderer& renderer, c
 
       // Calculate extraStartOffset for the first word on the line (i) (protect left margin)
       int extraStartOffset = 0;
-      if (j == i && !rubyTexts.empty() && i < rubyTexts.size() && !rubyTexts[i].empty() &&
+      if (j == static_cast<size_t>(i) && !rubyTexts.empty() && static_cast<size_t>(i) < rubyTexts.size() && !rubyTexts[i].empty() &&
           (wordStyles[i] & EpdFontFamily::RUBY_CONTINUE) == 0) {
         int groupWordCount = 1;
-        while (i + groupWordCount < totalWordCount &&
+        while (static_cast<size_t>(i + groupWordCount) < totalWordCount &&
                (wordStyles[i + groupWordCount] & EpdFontFamily::RUBY_CONTINUE) != 0) {
           groupWordCount++;
         }
@@ -790,7 +790,7 @@ std::vector<size_t> ParsedText::computeLineBreaks(const GfxRenderer& renderer, c
         }
       }
 
-      currlen += wordWidths[j] + gap + (j == i ? extraStartOffset : 0);
+      currlen += wordWidths[j] + gap + (j == static_cast<size_t>(i) ? extraStartOffset : 0);
 
       if (currlen > effectivePageWidth) {
         break;
